+++ pkg/provider/acceptance_cluster_replica_test.go
@@ -2,6 +2,7 @@ package provider
 
 import (
 	"database/sql"
+	"errors"
 	"fmt"
 	"testing"
 
@@ -161,7 +162,7 @@ func testAccCheckAllClusterReplicaDestroyed(s *terraform.State) error {
 		_, err := materialize.ScanClusterReplica(db, utils.ExtractId(r.Primary.ID))
 		if err == nil {
 			return fmt.Errorf("Cluster replica %v still exists", utils.ExtractId(r.Primary.ID))
-		} else if err != sql.ErrNoRows {
+		} else if !errors.Is(err, sql.ErrNoRows) {
 			return err
 		}
 	}