+++ pkg/provider/acceptance_role_test.go
@@ -2,6 +2,7 @@ package provider
 
 import (
 	"database/sql"
+	"errors"
 	"fmt"
 	"testing"
 
@@ -364,7 +365,7 @@ func testAccCheckAllRolesDestroyed(s *terraform.State) error {
 		_, err := materialize.ScanRole(db, utils.ExtractId(r.Primary.ID))
 		if err == nil {
 			return fmt.Errorf("role %v still exists", utils.ExtractId(r.Primary.ID))
-		} else if err != sql.ErrNoRows {
+		} else if !errors.Is(err, sql.ErrNoRows) {
 			return err
 		}
 	}