+++ pkg/provider/acceptance_source_table_postgres_test.go
@@ -2,6 +2,7 @@ package provider
 
 import (
 	"database/sql"
+	"errors"
 	"fmt"
 	"testing"
 
@@ -277,7 +278,7 @@ func testAccCheckAllSourceTablePostgresDestroyed(s *terraform.State) error {
 		_, err := materialize.ScanSourceTable(db, utils.ExtractId(r.Primary.ID))
 		if err == nil {
 			return fmt.Errorf("source table %v still exists", utils.ExtractId(r.Primary.ID))
-		} else if err != sql.ErrNoRows {
+		} else if !errors.Is(err, sql.ErrNoRows) {
 			return err
 		}
 	}