+++ pkg/provider/acceptance_view_test.go
@@ -2,6 +2,7 @@ package provider
 
 import (
 	"database/sql"
+	"errors"
 	"fmt"
 	"testing"
 
@@ -169,7 +170,7 @@ func testAccCheckAllViewsDestroyed(s *terraform.State) error {
 		_, err := materialize.ScanView(db, utils.ExtractId(r.Primary.ID))
 		if err == nil {
 			return fmt.Errorf("View %v still exists", utils.ExtractId(r.Primary.ID))
-		} else if err != sql.ErrNoRows {
+		} else if !errors.Is(err, sql.ErrNoRows) {
 			return err
 		}
 	}