+++ pkg/resources/resource_cluster.go
@@ -3,6 +3,7 @@ package resources
 import (
 	"context"
 	"database/sql"
+	"errors"
 	"fmt"
 	"log"
 	"regexp"
@@ -179,7 +180,7 @@ func clusterRead(ctx context.Context, d *schema.ResourceData, meta interface{}) 
 	useNameAsId := d.Get("identify_by_name").(bool)
 
 	s, err := materialize.ScanCluster(metaDb, value, idType == "name")
-	if err == sql.ErrNoRows {
+	if errors.Is(err, sql.ErrNoRows) {
 		d.SetId("")
 		return nil
 	} else if err != nil {