+++ pkg/resources/resource_index.go
@@ -3,7 +3,7 @@ package resources
 import (
 	"context"
 	"database/sql"
-	"log"
+	"errors"
 
 	"github.com/MaterializeInc/terraform-provider-materialize/pkg/materialize"
 	"github.com/MaterializeInc/terraform-provider-materialize/pkg/utils"
@@ -107,7 +107,7 @@ func indexRead(ctx context.Context, d *schema.ResourceData, meta interface{}) di
 		return diag.FromErr(err)
 	}
 	s, err := materialize.ScanIndex(metaDb, utils.ExtractId(i))
-	if err == sql.ErrNoRows {
+	if errors.Is(err, sql.ErrNoRows) {
 		d.SetId("")
 		return nil
 	} else if err != nil {
@@ -207,9 +207,7 @@ func indexCreate(ctx context.Context, d *schema.ResourceData, meta interface{}) 
 	// object comment
 	if v, ok := d.GetOk("comment"); ok {
 		if err := b.Comment(v.(string)); err != nil {
-			log.Printf("[DEBUG] resource failed comment, dropping object: %s", o.Name)
-			b.Drop()
-			return diag.FromErr(err)
+			return rollbackCreate(b, o, "comment", err)
 		}
 	}