+++ pkg/resources/resource_source_table_sqlserver.go
@@ -3,7 +3,7 @@ package resources
 import (
 	"context"
 	"database/sql"
-	"log"
+	"errors"
 
 	"github.com/MaterializeInc/terraform-provider-materialize/pkg/materialize"
 	"github.com/MaterializeInc/terraform-provider-materialize/pkg/utils"
@@ -123,9 +123,7 @@ func sourceTableSQLServerCreate(ctx context.Context, d *schema.ResourceData, met
 	if v, ok := d.GetOk("ownership_role"); ok {
 		ownership := materialize.NewOwnershipBuilder(metaDb, o)
 		if err := ownership.Alter(v.(string)); err != nil {
-			log.Printf("[DEBUG] resource failed ownership, dropping object: %s", o.Name)
-			b.Drop()
-			return diag.FromErr(err)
+			return rollbackCreate(b, o, "ownership", err)
 		}
 	}
 
@@ -133,9 +131,7 @@ func sourceTableSQLServerCreate(ctx context.Context, d *schema.ResourceData, met
 	if v, ok := d.GetOk("comment"); ok {
 		comment := materialize.NewCommentBuilder(metaDb, o)
 		if err := comment.Object(v.(string)); err != nil {
-			log.Printf("[DEBUG] resource failed comment, dropping object: %s", o.Name)
-			b.Drop()
-			return diag.FromErr(err)
+			return rollbackCreate(b, o, "comment", err)
 		}
 	}
 
@@ -199,7 +195,7 @@ func sourceTableSQLServerRead(ctx context.Context, d *schema.ResourceData, meta 
 	}
 
 	t, err := materialize.ScanSourceTableSQLServer(metaDb, utils.ExtractId(i))
-	if err == sql.ErrNoRows {
+	if errors.Is(err, sql.ErrNoRows) {
 		d.SetId("")
 		return nil
 	} else if err != nil {